Fix seven defects found reviewing the phase 0.1 fixes
Two of them cancelled the fix they were meant to deliver:

- main awaited the shutdown task after detecting that a background
  task had died, but that task is parked on ctrl_c()/SIGTERM and never
  resolves on its own, so the process hung instead of exiting non-zero.
  F1 detected the failure and drew no conclusion from it. Its result is
  now collected only if it has already finished, and dropped otherwise,
  which also surfaces a fatal signal-handler failure as a non-zero exit.
- GET / lived on dev.router, so gating that router behind the new safe
  default unmounted the homepage too and served a 404 in production.
  Public pages move to routers/pages.py, always mounted, and a test
  pins it: a public page never lives on a router we switch off.

Also: the backoff cast sat inside least() instead of wrapping it, so
it could overflow before the clamp applied (make_interval takes double
precision, so the cast is simply gone); the WORKER_ID fallback used the
pid alone, which is 1 in every container and reproduced the very zombie
reclaim failure F8 fixed; pg_isready probed the Unix socket that the
init-time temporary server also answers on; and the WORKER_ID test
mutated process environment under a SAFETY comment claiming a single
thread, so resolution is now a pure function tested without unsafe.

Verified against a real Postgres: ruff, ty, 13 backend tests, clippy,
5 unit and 12 integration Rust tests all green.

+++ worker/src/jobs/queue.rs
@@ -127,8 +127,10 @@ pub async fn fail_job(
         UPDATE job
         SET status       = (CASE WHEN attempts >= max_attempts THEN 'failed' ELSE 'pending' END)::job_status,
             error        = $3,
+            -- `make_interval(secs => …)` prend un double precision : pas de cast en int, qui
+            -- déborderait pour un max_attempts élevé avant même que least() ne plafonne.
             scheduled_at = now() + make_interval(
-                               secs => least(300, (5 * power(3, greatest(attempts - 1, 0)))::int)
+                               secs => least(300, 5 * power(3, greatest(attempts - 1, 0)))
                            ),
             locked_at    = NULL,
             locked_by    = NULL
